Check exitstatus on test cleanup (#1572)

+++ tests/prefect/conftest.py
@@ -26,8 +26,9 @@ if prefect:
             with prefect_test_harness():
                 yield
 
-    def pytest_sessionfinish():
-        from shutil import rmtree
+    def pytest_sessionfinish(exitstatus):
+        if exitstatus == 0:
+            from shutil import rmtree
 
-        rmtree(TEST_RESULTS_DIR, ignore_errors=True)
-        rmtree(TEST_SCRATCH_DIR, ignore_errors=True)
+            rmtree(TEST_RESULTS_DIR, ignore_errors=True)
+            rmtree(TEST_SCRATCH_DIR, ignore_errors=True)
